+++ changelog.md
@@ -13,6 +13,7 @@ This changelog helps developers and users to keep track of new features, fixes a
 * test: add tests for automatic migration of old configuration
 * refactor: reduce class visibility and add final to constructor assigned fields
 * feat: add logging for automatic plugin updates
+* feat: use transactions for writing Scrum Poker data (#32)
 
 # [3.5.0] - 2018-09-16